refactor : Lotto 에러 메시지 수정
